+++ examples/SimpleSynth/SimpleSynthProcessor.h
@@ -76,7 +76,7 @@ public:
     ~SimpleSynthProcessor(){};
 
     // Process method to handle audio processing
-    void process(float** inputs, float** outputs, std::size_t blockSize) override;
+    void process(double** inputs, double** outputs, std::size_t blockSize) override;
 
     // Set a parameter value
     void setParameter(int paramId, double value) override;
